New: Add en passant mechanic
Added the check_en_passant method to the Board,
along with the remove_en_passant method, which
mark squares' @en_passant_{color} properties
according to the player who last made a double
pawn move.

That square is then exploitable by the opponent
with a diagonal pawn move to take the piece even
though it is behind, as the rule dictates, but
only if it's done in the following move.

Also added en passant checks to the pawn class'
get_valid_squares method.