Fix: Double jump available even though piece is blocking
The pawn can no longer double jump over a piece
if that piece's in the first square ahead, blocking it.

Funnily, this came to my attention because I was
trying to find out why my stalemate wouldn't work;
turns out it was because the method calculating
all the possible moves that do not lead to a check
(find_all_available_dodges) was still returning
some strange pawn moves.